mgr/smb: move a handful of small functions to utils.py
Includes updating the files that were using them and even
a few test functions.
Amusingly, the `one` function was being tested as part of
test_handler.py demonstrating the confusion that not having
a utils.py and corresponding test file brought.
